+++ src/client.ts
@@ -1,9 +1,10 @@
-import http from 'http';
-
-import { IntegrationProviderAuthenticationError } from '@jupiterone/integration-sdk-core';
-
 import { IntegrationConfig } from './config';
-import { AcmeUser, AcmeGroup } from './types';
+import { LdapClient, LdapTestAdapter, LdapTSAdapter } from './ldap';
+import {
+  ActiveDirectoryUser,
+  ActiveDirectoryGroup,
+  ActiveDirectoryComputer,
+} from './types';
 
 export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
 
@@ -16,41 +17,13 @@ export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
  * resources.
  */
 export class APIClient {
-  constructor(readonly config: IntegrationConfig) {}
+  constructor(private readonly client: LdapClient) {}
 
+  /**
+   * Verifies authentication by doing a bind action.
+   */
   public async verifyAuthentication(): Promise<void> {
-    // TODO make the most light-weight request possible to validate
-    // authentication works with the provided credentials, throw an err if
-    // authentication fails
-    const request = new Promise<void>((resolve, reject) => {
-      http.get(
-        {
-          hostname: 'localhost',
-          port: 443,
-          path: '/api/v1/some/endpoint?limit=1',
-          agent: false,
-          timeout: 10,
-        },
-        (res) => {
-          if (res.statusCode !== 200) {
-            reject(new Error('Provider authentication failed'));
-          } else {
-            resolve();
-          }
-        },
-      );
-    });
-
-    try {
-      await request;
-    } catch (err) {
-      throw new IntegrationProviderAuthenticationError({
-        cause: err,
-        endpoint: 'https://localhost/api/v1/some/endpoint?limit=1',
-        status: err.status,
-        statusText: err.statusText,
-      });
-    }
+    return this.client.verifyAuthentication();
   }
 
   /**
@@ -59,26 +32,11 @@ export class APIClient {
    * @param iteratee receives each resource to produce entities/relationships
    */
   public async iterateUsers(
-    iteratee: ResourceIteratee<AcmeUser>,
+    iteratee: ResourceIteratee<ActiveDirectoryUser>,
   ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
-
-    const users: AcmeUser[] = [
-      {
-        id: 'acme-user-1',
-        name: 'User One',
-      },
-      {
-        id: 'acme-user-2',
-        name: 'User Two',
-      },
-    ];
+    const users: ActiveDirectoryUser[] = await this.client.search(
+      'objectCategory=User',
+    );
 
     for (const user of users) {
       await iteratee(user);
@@ -91,34 +49,46 @@ export class APIClient {
    * @param iteratee receives each resource to produce entities/relationships
    */
   public async iterateGroups(
-    iteratee: ResourceIteratee<AcmeGroup>,
+    iteratee: ResourceIteratee<ActiveDirectoryGroup>,
   ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
-
-    const groups: AcmeGroup[] = [
-      {
-        id: 'acme-group-1',
-        name: 'Group One',
-        users: [
-          {
-            id: 'acme-user-1',
-          },
-        ],
-      },
-    ];
+    const groups: ActiveDirectoryGroup[] = await this.client.search(
+      'objectClass=Group',
+    );
 
     for (const group of groups) {
       await iteratee(group);
     }
   }
+
+  /**
+   * Iterates each device resource in the provider.
+   *
+   * @param iteratee receives each resource to produce entities/relationships
+   */
+  public async iterateDevices(
+    iteratee: ResourceIteratee<ActiveDirectoryComputer>,
+  ): Promise<void> {
+    const devices: ActiveDirectoryComputer[] = await this.client.search(
+      'objectClass=Computer',
+    );
+
+    for (const device of devices) {
+      await iteratee(device);
+    }
+  }
 }
 
 export function createAPIClient(config: IntegrationConfig): APIClient {
-  return new APIClient(config);
+  if (process.env.NODE_ENV === 'test') {
+    return new APIClient(new LdapTestAdapter());
+  }
+
+  return new APIClient(
+    new LdapTSAdapter({
+      domain: config.clientDomain,
+      url: config.clientUrl,
+      username: config.clientUsername,
+      password: config.clientPassword,
+    }),
+  );
 }