+++ src/steps/access/index.ts
@@ -8,15 +8,18 @@ import {
 
 import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
-import { AcmeGroup } from '../../types';
+import { ActiveDirectoryGroup, ActiveDirectoryUser } from '../../types';
 import { ACCOUNT_ENTITY_KEY } from '../account';
 import { Entities, Steps, Relationships } from '../constants';
 import {
   createAccountGroupRelationship,
   createAccountUserRelationship,
+  createDeviceEntity,
   createGroupEntity,
+  createGroupGroupRelationship,
   createGroupUserRelationship,
   createUserEntity,
+  getGroupId,
 } from './converter';
 
 export async function fetchUsers({
@@ -29,6 +32,7 @@ export async function fetchUsers({
 
   await apiClient.iterateUsers(async (user) => {
     const userEntity = await jobState.addEntity(createUserEntity(user));
+
     await jobState.addRelationship(
       createAccountUserRelationship(accountEntity, userEntity),
     );
@@ -45,8 +49,26 @@ export async function fetchGroups({
 
   await apiClient.iterateGroups(async (group) => {
     const groupEntity = await jobState.addEntity(createGroupEntity(group));
+
     await jobState.addRelationship(
       createAccountGroupRelationship(accountEntity, groupEntity),
+    );
+  });
+}
+
+export async function fetchDevices({
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const apiClient = createAPIClient(instance.config);
+
+  const accountEntity = (await jobState.getData(ACCOUNT_ENTITY_KEY)) as Entity;
+
+  await apiClient.iterateDevices(async (computer) => {
+    const deviceEntity = await jobState.addEntity(createDeviceEntity(computer));
+
+    await jobState.addRelationship(
+      createAccountGroupRelationship(accountEntity, deviceEntity),
     );
   });
 }
@@ -56,29 +78,80 @@ export async function buildGroupUserRelationships({
   logger,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
   await jobState.iterateEntities(
+    { _type: Entities.USER._type },
+    async (userEntity) => {
+      const user = getRawData<ActiveDirectoryUser>(userEntity);
+
+      if (!user) {
+        logger.warn(
+          { _key: userEntity._key },
+          'Could not get raw data for user entity',
+        );
+
+        return;
+      }
+
+      if (!user.memberOf) {
+        return;
+      }
+
+      const userGroups =
+        typeof user.memberOf === 'string' ? [user.memberOf] : user.memberOf;
+
+      for (const groupId of userGroups) {
+        const groupEntity = await jobState.findEntity(getGroupId(groupId));
+
+        if (!groupEntity) {
+          throw new IntegrationMissingKeyError(
+            `Expected group with key to exist (key=${groupId})`,
+          );
+        }
+
+        await jobState.addRelationship(
+          createGroupUserRelationship(groupEntity, userEntity),
+        );
+      }
+    },
+  );
+}
+
+export async function buildGroupGroupRelationships({
+  jobState,
+  logger,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  await jobState.iterateEntities(
     { _type: Entities.GROUP._type },
     async (groupEntity) => {
-      const group = getRawData<AcmeGroup>(groupEntity);
+      const group = getRawData<ActiveDirectoryGroup>(groupEntity);
 
       if (!group) {
         logger.warn(
           { _key: groupEntity._key },
           'Could not get raw data for group entity',
         );
+
         return;
       }
 
-      for (const user of group.users || []) {
-        const userEntity = await jobState.findEntity(user.id);
+      if (!group.memberOf) {
+        return;
+      }
 
-        if (!userEntity) {
+      const groupGroups =
+        typeof group.memberOf === 'string' ? [group.memberOf] : group.memberOf;
+
+      for (const memberOf of groupGroups) {
+        const sourceGroupId = getGroupId(memberOf);
+        const sourceGroupEntity = await jobState.findEntity(sourceGroupId);
+
+        if (!sourceGroupEntity) {
           throw new IntegrationMissingKeyError(
-            `Expected user with key to exist (key=${user.id})`,
+            `Expected group with key to exist (key=${sourceGroupId})`,
           );
         }
 
         await jobState.addRelationship(
-          createGroupUserRelationship(groupEntity, userEntity),
+          createGroupGroupRelationship(sourceGroupEntity, groupEntity),
         );
       }
     },
@@ -103,11 +176,27 @@ export const accessSteps: IntegrationStep<IntegrationConfig>[] = [
     executionHandler: fetchGroups,
   },
   {
+    id: Steps.DEVICES,
+    name: 'Fetch Devices',
+    entities: [Entities.DEVICE],
+    relationships: [Relationships.ACCOUNT_HAS_DEVICE],
+    dependsOn: [Steps.ACCOUNT],
+    executionHandler: fetchDevices,
+  },
+  {
     id: Steps.GROUP_USER_RELATIONSHIPS,
     name: 'Build Group -> User Relationships',
     entities: [],
     relationships: [Relationships.GROUP_HAS_USER],
     dependsOn: [Steps.GROUPS, Steps.USERS],
     executionHandler: buildGroupUserRelationships,
+  },
+  {
+    id: Steps.GROUP_GROUP_RELATIONSHIPS,
+    name: 'Build Group -> Group Relationships',
+    entities: [],
+    relationships: [Relationships.GROUP_HAS_GROUP],
+    dependsOn: [Steps.GROUPS],
+    executionHandler: buildGroupGroupRelationships,
   },
 ];