+++ src/types.ts
@@ -1,25 +1,33 @@
-// Providers often supply types with their API libraries.
-
-export interface AcmeUser {
-  id: string;
+export interface ActiveDirectoryUser {
+  dn: string;
+  cn: string;
+  objectGUID: string;
   name: string;
+  memberOf?: string | string[];
+  description: string;
+  whenCreated: string;
+  whenChanged: string;
 }
 
-export interface AcmeGroup {
-  id: string;
+export interface ActiveDirectoryGroup {
+  dn: string;
+  cn: string;
+  objectGUID: string;
   name: string;
-  users?: Pick<AcmeUser, 'id'>[];
+  description: string;
+  whenCreated: string;
+  whenChanged: string;
+  memberOf?: string | string[];
 }
 
-// Those can be useful to a degree, but often they're just full of optional
-// values. Understanding the response data may be more reliably accomplished by
-// reviewing the API response recordings produced by testing the wrapper client
-// (./client.ts). However, when there are no types provided, it is necessary to define
-// opaque types for each resource, to communicate the records that are expected
-// to come from an endpoint and are provided to iterating functions.
-
-/*
-import { Opaque } from 'type-fest';
-export type AcmeUser = Opaque<any, 'AcmeUser'>;
-export type AcmeGroup = Opaque<any, 'AcmeGroup'>;
-*/
+export interface ActiveDirectoryComputer {
+  dn: string;
+  cn: string;
+  objectGUID: string;
+  name: string;
+  instanceType: string;
+  operatingSystem: string;
+  operatingSystemVersion: string;
+  whenCreated: string;
+  whenChanged: string;
+}