+++ src/validateInvocation.test.ts
@@ -21,7 +21,7 @@ describe('#validateInvocation', () => {
     });
 
     await expect(validateInvocation(executionContext)).rejects.toThrow(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of {clientUrl, clientUsername, clientPassword, clientDomain}',
     );
   });
 
@@ -54,10 +54,10 @@ describe('#validateInvocation', () => {
      * error messaging is expected and clear to end-users
      */
     describe('invalid user credentials', () => {
-      test.skip('should throw if clientId is invalid', async () => {
+      test.skip('should throw if clientUrl is invalid', async () => {
         recording = setupProjectRecording({
           directory: __dirname,
-          name: 'client-id-auth-error',
+          name: 'client-url-auth-error',
           // Many authorization failures will return non-200 responses
           // and `recordFailedRequest: true` is needed to capture these responses
           options: {
@@ -67,8 +67,8 @@ describe('#validateInvocation', () => {
 
         const executionContext = createMockExecutionContext({
           instanceConfig: {
-            clientId: 'INVALID',
-            clientSecret: integrationConfig.clientSecret,
+            ...integrationConfig,
+            clientUrl: 'INVALID',
           },
         });
 
@@ -79,10 +79,12 @@ describe('#validateInvocation', () => {
         );
       });
 
-      test.skip('should throw if clientSecret is invalid', async () => {
+      test.skip('should throw if clientDomain is invalid', async () => {
         recording = setupProjectRecording({
           directory: __dirname,
-          name: 'client-secret-auth-error',
+          name: 'client-domain-auth-error',
+          // Many authorization failures will return non-200 responses
+          // and `recordFailedRequest: true` is needed to capture these responses
           options: {
             recordFailedRequests: true,
           },
@@ -90,11 +92,66 @@ describe('#validateInvocation', () => {
 
         const executionContext = createMockExecutionContext({
           instanceConfig: {
-            clientId: integrationConfig.clientSecret,
-            clientSecret: 'INVALID',
+            ...integrationConfig,
+            clientUrl: 'ldap://localhost',
+            clientDomain: 'INVALID',
           },
         });
 
+        // tests validate that invalid configurations throw an error
+        // with an appropriate and expected message.
+        await expect(validateInvocation(executionContext)).rejects.toThrow(
+          'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
+        );
+      });
+
+      test.skip('should throw if clientUsername is invalid', async () => {
+        recording = setupProjectRecording({
+          directory: __dirname,
+          name: 'client-username-auth-error',
+          // Many authorization failures will return non-200 responses
+          // and `recordFailedRequest: true` is needed to capture these responses
+          options: {
+            recordFailedRequests: true,
+          },
+        });
+
+        const executionContext = createMockExecutionContext({
+          instanceConfig: {
+            ...integrationConfig,
+            clientUrl: 'ldap://localhost',
+            clientUsername: 'INVALID',
+          },
+        });
+
+        // tests validate that invalid configurations throw an error
+        // with an appropriate and expected message.
+        await expect(validateInvocation(executionContext)).rejects.toThrow(
+          'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
+        );
+      });
+
+      test.skip('should throw if clientPassword is invalid', async () => {
+        recording = setupProjectRecording({
+          directory: __dirname,
+          name: 'client-password-auth-error',
+          // Many authorization failures will return non-200 responses
+          // and `recordFailedRequest: true` is needed to capture these responses
+          options: {
+            recordFailedRequests: true,
+          },
+        });
+
+        const executionContext = createMockExecutionContext({
+          instanceConfig: {
+            ...integrationConfig,
+            clientUrl: 'ldap://localhost',
+            clientPassword: 'INVALID',
+          },
+        });
+
+        // tests validate that invalid configurations throw an error
+        // with an appropriate and expected message.
         await expect(validateInvocation(executionContext)).rejects.toThrow(
           'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
         );